+++ Lain/OptionsForm.Designer.cs
@@ -41,12 +41,11 @@
             this.checkBox3 = new System.Windows.Forms.CheckBox();
             this.txtTimer = new System.Windows.Forms.TextBox();
             this.label1 = new System.Windows.Forms.Label();
-            this.btnBackup = new System.Windows.Forms.Button();
             this.sizePanel = new System.Windows.Forms.Panel();
-            this.rSmall = new System.Windows.Forms.RadioButton();
-            this.rNormal = new System.Windows.Forms.RadioButton();
-            this.rLarge = new System.Windows.Forms.RadioButton();
             this.label2 = new System.Windows.Forms.Label();
+            this.rLarge = new System.Windows.Forms.RadioButton();
+            this.rNormal = new System.Windows.Forms.RadioButton();
+            this.rSmall = new System.Windows.Forms.RadioButton();
             this.checkBox4 = new System.Windows.Forms.CheckBox();
             this.sizePanel.SuspendLayout();
             this.SuspendLayout();
@@ -55,9 +54,10 @@
             // 
             this.checkBox1.AutoSize = true;
             this.checkBox1.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.checkBox1.Location = new System.Drawing.Point(17, 345);
+            this.checkBox1.Location = new System.Drawing.Point(14, 276);
+            this.checkBox1.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.checkBox1.Name = "checkBox1";
-            this.checkBox1.Size = new System.Drawing.Size(256, 32);
+            this.checkBox1.Size = new System.Drawing.Size(206, 25);
             this.checkBox1.TabIndex = 0;
             this.checkBox1.Text = "Always ask for password";
             this.checkBox1.UseVisualStyleBackColor = true;
@@ -68,9 +68,10 @@
             this.minimaltheme.AutoSize = true;
             this.minimaltheme.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.minimaltheme.ForeColor = System.Drawing.Color.Gray;
-            this.minimaltheme.Location = new System.Drawing.Point(153, 126);
+            this.minimaltheme.Location = new System.Drawing.Point(122, 101);
+            this.minimaltheme.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.minimaltheme.Name = "minimaltheme";
-            this.minimaltheme.Size = new System.Drawing.Size(106, 32);
+            this.minimaltheme.Size = new System.Drawing.Size(86, 25);
             this.minimaltheme.TabIndex = 76;
             this.minimaltheme.Text = "Minimal";
             this.minimaltheme.UseVisualStyleBackColor = true;
@@ -81,9 +82,10 @@
             this.carameltheme.AutoSize = true;
             this.carameltheme.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.carameltheme.ForeColor = System.Drawing.Color.DarkOrange;
-            this.carameltheme.Location = new System.Drawing.Point(153, 50);
+            this.carameltheme.Location = new System.Drawing.Point(122, 40);
+            this.carameltheme.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.carameltheme.Name = "carameltheme";
-            this.carameltheme.Size = new System.Drawing.Size(106, 32);
+            this.carameltheme.Size = new System.Drawing.Size(87, 25);
             this.carameltheme.TabIndex = 75;
             this.carameltheme.Text = "Caramel";
             this.carameltheme.UseVisualStyleBackColor = true;
@@ -94,9 +96,10 @@
             this.limetheme.AutoSize = true;
             this.limetheme.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.limetheme.ForeColor = System.Drawing.Color.LimeGreen;
-            this.limetheme.Location = new System.Drawing.Point(153, 88);
+            this.limetheme.Location = new System.Drawing.Point(122, 70);
+            this.limetheme.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.limetheme.Name = "limetheme";
-            this.limetheme.Size = new System.Drawing.Size(77, 32);
+            this.limetheme.Size = new System.Drawing.Size(63, 25);
             this.limetheme.TabIndex = 74;
             this.limetheme.Text = "Lime";
             this.limetheme.UseVisualStyleBackColor = true;
@@ -107,9 +110,10 @@
             this.magmatheme.AutoSize = true;
             this.magmatheme.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.magmatheme.ForeColor = System.Drawing.Color.Tomato;
-            this.magmatheme.Location = new System.Drawing.Point(37, 88);
+            this.magmatheme.Location = new System.Drawing.Point(30, 70);
+            this.magmatheme.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.magmatheme.Name = "magmatheme";
-            this.magmatheme.Size = new System.Drawing.Size(101, 32);
+            this.magmatheme.Size = new System.Drawing.Size(83, 25);
             this.magmatheme.TabIndex = 73;
             this.magmatheme.Text = "Magma";
             this.magmatheme.UseVisualStyleBackColor = true;
@@ -121,9 +125,10 @@
             this.oceantheme.Checked = true;
             this.oceantheme.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.oceantheme.ForeColor = System.Drawing.Color.DodgerBlue;
-            this.oceantheme.Location = new System.Drawing.Point(37, 50);
+            this.oceantheme.Location = new System.Drawing.Point(30, 40);
+            this.oceantheme.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.oceantheme.Name = "oceantheme";
-            this.oceantheme.Size = new System.Drawing.Size(90, 32);
+            this.oceantheme.Size = new System.Drawing.Size(74, 25);
             this.oceantheme.TabIndex = 72;
             this.oceantheme.TabStop = true;
             this.oceantheme.Text = "Ocean";
@@ -135,9 +140,10 @@
             this.zergtheme.AutoSize = true;
             this.zergtheme.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.zergtheme.ForeColor = System.Drawing.Color.MediumOrchid;
-            this.zergtheme.Location = new System.Drawing.Point(37, 126);
+            this.zergtheme.Location = new System.Drawing.Point(30, 101);
+            this.zergtheme.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.zergtheme.Name = "zergtheme";
-            this.zergtheme.Size = new System.Drawing.Size(75, 32);
+            this.zergtheme.Size = new System.Drawing.Size(62, 25);
             this.zergtheme.TabIndex = 71;
             this.zergtheme.Text = "Zerg";
             this.zergtheme.UseVisualStyleBackColor = true;
@@ -148,9 +154,10 @@
             this.label27.AutoSize = true;
             this.label27.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.label27.ForeColor = System.Drawing.Color.DodgerBlue;
-            this.label27.Location = new System.Drawing.Point(12, 9);
+            this.label27.Location = new System.Drawing.Point(10, 7);
+            this.label27.Margin = new System.Windows.Forms.Padding(2, 0, 2, 0);
             this.label27.Name = "label27";
-            this.label27.Size = new System.Drawing.Size(192, 28);
+            this.label27.Size = new System.Drawing.Size(153, 21);
             this.label27.TabIndex = 70;
             this.label27.Tag = "themeable";
             this.label27.Text = "Choose your theme";
@@ -166,9 +173,10 @@
             this.btnOk.FlatStyle = System.Windows.Forms.FlatStyle.Flat;
             this.btnOk.Font = new System.Drawing.Font("Segoe UI Semibold", 9F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.btnOk.ForeColor = System.Drawing.Color.White;
-            this.btnOk.Location = new System.Drawing.Point(269, 521);
+            this.btnOk.Location = new System.Drawing.Point(215, 417);
+            this.btnOk.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.btnOk.Name = "btnOk";
-            this.btnOk.Size = new System.Drawing.Size(89, 39);
+            this.btnOk.Size = new System.Drawing.Size(71, 31);
             this.btnOk.TabIndex = 77;
             this.btnOk.Tag = "themeable";
             this.btnOk.Text = "OK";
@@ -179,9 +187,10 @@
             // 
             this.checkBox2.AutoSize = true;
             this.checkBox2.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.checkBox2.Location = new System.Drawing.Point(17, 421);
+            this.checkBox2.Location = new System.Drawing.Point(14, 337);
+            this.checkBox2.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.checkBox2.Name = "checkBox2";
-            this.checkBox2.Size = new System.Drawing.Size(144, 32);
+            this.checkBox2.Size = new System.Drawing.Size(116, 25);
             this.checkBox2.TabIndex = 78;
             this.checkBox2.Text = "Auto lock in";
             this.checkBox2.UseVisualStyleBackColor = true;
@@ -191,9 +200,10 @@
             // 
             this.checkBox3.AutoSize = true;
             this.checkBox3.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.checkBox3.Location = new System.Drawing.Point(17, 459);
+            this.checkBox3.Location = new System.Drawing.Point(14, 367);
+            this.checkBox3.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.checkBox3.Name = "checkBox3";
-            this.checkBox3.Size = new System.Drawing.Size(254, 32);
+            this.checkBox3.Size = new System.Drawing.Size(203, 25);
             this.checkBox3.TabIndex = 79;
             this.checkBox3.Text = "Start Lain with Windows";
             this.checkBox3.UseVisualStyleBackColor = true;
@@ -206,10 +216,11 @@
             this.txtTimer.Enabled = false;
             this.txtTimer.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.txtTimer.ForeColor = System.Drawing.Color.White;
-            this.txtTimer.Location = new System.Drawing.Point(167, 420);
+            this.txtTimer.Location = new System.Drawing.Point(134, 336);
+            this.txtTimer.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.txtTimer.Name = "txtTimer";
             this.txtTimer.ShortcutsEnabled = false;
-            this.txtTimer.Size = new System.Drawing.Size(62, 34);
+            this.txtTimer.Size = new System.Drawing.Size(50, 29);
             this.txtTimer.TabIndex = 80;
             this.txtTimer.Text = "2";
             this.txtTimer.TextChanged += new System.EventHandler(this.txtTimer_TextChanged);
@@ -219,30 +230,12 @@
             // 
             this.label1.AutoSize = true;
             this.label1.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.label1.Location = new System.Drawing.Point(235, 422);
+            this.label1.Location = new System.Drawing.Point(188, 338);
+            this.label1.Margin = new System.Windows.Forms.Padding(2, 0, 2, 0);
             this.label1.Name = "label1";
-            this.label1.Size = new System.Drawing.Size(86, 28);
+            this.label1.Size = new System.Drawing.Size(68, 21);
             this.label1.TabIndex = 81;
             this.label1.Text = "minutes";
-            // 
-            // btnBackup
-            // 
-            this.btnBackup.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
-            this.btnBackup.BackColor = System.Drawing.Color.DodgerBlue;
-            this.btnBackup.FlatAppearance.BorderColor = System.Drawing.Color.DodgerBlue;
-            this.btnBackup.FlatAppearance.MouseDownBackColor = System.Drawing.Color.RoyalBlue;
-            this.btnBackup.FlatAppearance.MouseOverBackColor = System.Drawing.Color.RoyalBlue;
-            this.btnBackup.FlatStyle = System.Windows.Forms.FlatStyle.Flat;
-            this.btnBackup.Font = new System.Drawing.Font("Segoe UI Semibold", 9F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.btnBackup.ForeColor = System.Drawing.Color.White;
-            this.btnBackup.Location = new System.Drawing.Point(12, 521);
-            this.btnBackup.Name = "btnBackup";
-            this.btnBackup.Size = new System.Drawing.Size(137, 39);
-            this.btnBackup.TabIndex = 82;
-            this.btnBackup.Tag = "themeable";
-            this.btnBackup.Text = "Create backup";
-            this.btnBackup.UseVisualStyleBackColor = false;
-            this.btnBackup.Click += new System.EventHandler(this.button1_Click);
             // 
             // sizePanel
             // 
@@ -250,49 +243,11 @@
             this.sizePanel.Controls.Add(this.rLarge);
             this.sizePanel.Controls.Add(this.rNormal);
             this.sizePanel.Controls.Add(this.rSmall);
-            this.sizePanel.Location = new System.Drawing.Point(12, 174);
+            this.sizePanel.Location = new System.Drawing.Point(10, 139);
+            this.sizePanel.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.sizePanel.Name = "sizePanel";
-            this.sizePanel.Size = new System.Drawing.Size(276, 165);
+            this.sizePanel.Size = new System.Drawing.Size(221, 132);
             this.sizePanel.TabIndex = 83;
-            // 
-            // rSmall
-            // 
-            this.rSmall.AutoSize = true;
-            this.rSmall.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.rSmall.ForeColor = System.Drawing.Color.White;
-            this.rSmall.Location = new System.Drawing.Point(25, 40);
-            this.rSmall.Name = "rSmall";
-            this.rSmall.Size = new System.Drawing.Size(82, 32);
-            this.rSmall.TabIndex = 77;
-            this.rSmall.Text = "Small";
-            this.rSmall.UseVisualStyleBackColor = true;
-            this.rSmall.CheckedChanged += new System.EventHandler(this.rSmall_CheckedChanged);
-            // 
-            // rNormal
-            // 
-            this.rNormal.AutoSize = true;
-            this.rNormal.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.rNormal.ForeColor = System.Drawing.Color.White;
-            this.rNormal.Location = new System.Drawing.Point(25, 78);
-            this.rNormal.Name = "rNormal";
-            this.rNormal.Size = new System.Drawing.Size(100, 32);
-            this.rNormal.TabIndex = 78;
-            this.rNormal.Text = "Normal";
-            this.rNormal.UseVisualStyleBackColor = true;
-            this.rNormal.CheckedChanged += new System.EventHandler(this.rNormal_CheckedChanged);
-            // 
-            // rLarge
-            // 
-            this.rLarge.AutoSize = true;
-            this.rLarge.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.rLarge.ForeColor = System.Drawing.Color.White;
-            this.rLarge.Location = new System.Drawing.Point(25, 116);
-            this.rLarge.Name = "rLarge";
-            this.rLarge.Size = new System.Drawing.Size(83, 32);
-            this.rLarge.TabIndex = 79;
-            this.rLarge.Text = "Large";
-            this.rLarge.UseVisualStyleBackColor = true;
-            this.rLarge.CheckedChanged += new System.EventHandler(this.rLarge_CheckedChanged);
             // 
             // label2
             // 
@@ -300,19 +255,63 @@
             this.label2.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.label2.ForeColor = System.Drawing.Color.DodgerBlue;
             this.label2.Location = new System.Drawing.Point(0, 0);
+            this.label2.Margin = new System.Windows.Forms.Padding(2, 0, 2, 0);
             this.label2.Name = "label2";
-            this.label2.Size = new System.Drawing.Size(211, 28);
+            this.label2.Size = new System.Drawing.Size(168, 21);
             this.label2.TabIndex = 84;
             this.label2.Tag = "themeable";
             this.label2.Text = "Choose your font size";
+            // 
+            // rLarge
+            // 
+            this.rLarge.AutoSize = true;
+            this.rLarge.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+            this.rLarge.ForeColor = System.Drawing.Color.White;
+            this.rLarge.Location = new System.Drawing.Point(20, 93);
+            this.rLarge.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
+            this.rLarge.Name = "rLarge";
+            this.rLarge.Size = new System.Drawing.Size(69, 25);
+            this.rLarge.TabIndex = 79;
+            this.rLarge.Text = "Large";
+            this.rLarge.UseVisualStyleBackColor = true;
+            this.rLarge.CheckedChanged += new System.EventHandler(this.rLarge_CheckedChanged);
+            // 
+            // rNormal
+            // 
+            this.rNormal.AutoSize = true;
+            this.rNormal.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+            this.rNormal.ForeColor = System.Drawing.Color.White;
+            this.rNormal.Location = new System.Drawing.Point(20, 62);
+            this.rNormal.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
+            this.rNormal.Name = "rNormal";
+            this.rNormal.Size = new System.Drawing.Size(82, 25);
+            this.rNormal.TabIndex = 78;
+            this.rNormal.Text = "Normal";
+            this.rNormal.UseVisualStyleBackColor = true;
+            this.rNormal.CheckedChanged += new System.EventHandler(this.rNormal_CheckedChanged);
+            // 
+            // rSmall
+            // 
+            this.rSmall.AutoSize = true;
+            this.rSmall.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+            this.rSmall.ForeColor = System.Drawing.Color.White;
+            this.rSmall.Location = new System.Drawing.Point(20, 32);
+            this.rSmall.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
+            this.rSmall.Name = "rSmall";
+            this.rSmall.Size = new System.Drawing.Size(67, 25);
+            this.rSmall.TabIndex = 77;
+            this.rSmall.Text = "Small";
+            this.rSmall.UseVisualStyleBackColor = true;
+            this.rSmall.CheckedChanged += new System.EventHandler(this.rSmall_CheckedChanged);
             // 
             // checkBox4
             // 
             this.checkBox4.AutoSize = true;
             this.checkBox4.Font = new System.Drawing.Font("Segoe UI Semibold", 12F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.checkBox4.Location = new System.Drawing.Point(17, 383);
+            this.checkBox4.Location = new System.Drawing.Point(14, 306);
+            this.checkBox4.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.checkBox4.Name = "checkBox4";
-            this.checkBox4.Size = new System.Drawing.Size(178, 32);
+            this.checkBox4.Size = new System.Drawing.Size(145, 25);
             this.checkBox4.TabIndex = 84;
             this.checkBox4.Text = "Hide passwords";
             this.checkBox4.UseVisualStyleBackColor = true;
@@ -321,14 +320,13 @@
             // OptionsForm
             // 
             this.AcceptButton = this.btnOk;
-            this.AutoScaleDimensions = new System.Drawing.SizeF(120F, 120F);
+            this.AutoScaleDimensions = new System.Drawing.SizeF(96F, 96F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Dpi;
             this.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(20)))), ((int)(((byte)(20)))), ((int)(((byte)(20)))));
             this.CancelButton = this.btnOk;
-            this.ClientSize = new System.Drawing.Size(370, 572);
+            this.ClientSize = new System.Drawing.Size(296, 458);
             this.Controls.Add(this.checkBox4);
             this.Controls.Add(this.sizePanel);
-            this.Controls.Add(this.btnBackup);
             this.Controls.Add(this.label1);
             this.Controls.Add(this.txtTimer);
             this.Controls.Add(this.checkBox3);
@@ -345,6 +343,7 @@
             this.Font = new System.Drawing.Font("Segoe UI Semibold", 9F, System.Drawing.FontStyle.Bold, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.ForeColor = System.Drawing.Color.White;
             this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.FixedSingle;
+            this.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.MaximizeBox = false;
             this.MinimizeBox = false;
             this.Name = "OptionsForm";
@@ -375,7 +374,6 @@
         private System.Windows.Forms.CheckBox checkBox3;
         private System.Windows.Forms.TextBox txtTimer;
         private System.Windows.Forms.Label label1;
-        private System.Windows.Forms.Button btnBackup;
         private System.Windows.Forms.Panel sizePanel;
         private System.Windows.Forms.Label label2;
         private System.Windows.Forms.RadioButton rLarge;